requirements.txt
# EvoAlpha Dependencies
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
supabase>=2.0.0
openai>=1.0.0
tabulate>=0.9.0
